+++ src/main/java/com/pro/list_tick/shopping_list/dto/ExpenseRequestUpdateDTO.java
@@ -1,5 +1,7 @@
 package com.pro.list_tick.shopping_list.dto;
 
+import java.math.BigDecimal;
+
 import com.pro.list_tick.shopping_list.model.CurrencyCode;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
@@ -12,7 +14,7 @@ public record ExpenseRequestUpdateDTO(
 
     @NotNull(message = "Amount cannot be null")
     @Positive(message = "Amount cannot be negative")
-    Double amount,
+    BigDecimal amount,
 
     @NotBlank(message = "Currency cannot be blank")
     @Size(min = 3, max = 3, message = "Currency has to be 3 characters long")
